Split decide-protocol into internal and external
